Export fix completetion.
The fix 3 from nekogaijin was working but I found another related issue,
in the exportFrames1 fn: modulo was not used so some extra time was not
filled with cloned frames.
Now, exported FPS can be smaller, iqual or greater than editor FPS.
I checked all the cases and it worked fine.