Improve movie export dialog
In addition to the main progress bar, a second progress bar and a
label have been added to display subtasks of the exporting. Not only
does this better show progress, but it also makes reports about
freezing/crashing during export more useful (ex. "it says Generating
Palette and then crashes").

The ffmpeg execution code was modified to detect the progress
of the operation. This also fixes a hidden serious bug with movie
exporting: waitForFinished has a timeout of 30 seconds by default,
so any movie that takes longer than that to export would timeout
and fail. This should hopeful fix some of the exporting errors
we've been seeing. waitForReadyRead also has a timeout (default 30s
I think), but this isn't an issue because FFmpeg will write updates
and flush every second or so, completing waitForReadyRead well
before the timeout.

This commit also fixes another issue, where it was impossible to
cancel during the movie export operation because processEvents was
only every being called with QEventLoop::ExcludeUserInputEvents,
which prevented the cancel button event from registering until
after the export is done!