Remove Show Tab Bar from the View menu on mac
This was done by integrating a small amount of native objective-c
code for mac since Qt has no way to do this at the moment.

This also enables us to easily add support for more native mac
features in the future. Mouse coalescing has been implemented
in this same way, but is not currently being used. See PR for
details.