#1112 - Bring back relative layer transparency
This brings back relative layer transparency with a slight enhancement or similarity to the real world.
The top layer is at the bottom of the visibility stack, meaning that once the top is selected, other layers aren't shown. As soon as you move down the stack (or add more papers to the light board) then the drawings behind will become more and more opaque.