Improve visibility cycling when clicking multiple times
Because the parent implementation of mouseDoubleClickEvent was
never being called, the second mouse event never fires, causing
some mouse presses to not change the visibility. This wasn't
a big issue with most of the other timeline functions because
clicking twice in the same place usually has the same effect as
clicking once or a double click action.