Fix playback audio sync issues
These changes will hopefully completely fix the audio-sync issues
during in-app playback. The first change was to keep all the audio
in memory all the time, not exactly efficient, but eliminates the
issues associated with hard drive seeks etc. The second change was
rearranging when playbackmanager plays the sound. There was a 1
frame offset because it was playing the sound of the frame right
before switch to the next frame, and there was a 1 frame of silence
at the beginning of playback because playsound was not called until
the first scrub.