Fix assertion triggered when opening projects sometimes
If you open a project or create a new project, and the current
layer index is out of bounds on the new project, it will try and
get the current layer in a timeline update a fail. This timeline
update is caused by scrubbing to the last frame in the opening
project, so the simplest solution is to simply set the layer index
before the frame position.