Improve center overlay rendering
Depending on the zoom level, the center overlay sometimes would
not show lines right in the center, or would only show half of the
lines because it is dashed. By drawing four lines from the center
instead of two through the center, it seems there will always be
lines at the center, even when zoomed out or in.