Move quick size handling to the tools
In addition to be being a more logical location for this code, it
also allows for more customization of the quick sizing to each
tools specific needs. For example, the bucket tool uses it for
adjusting the fill expansion, but that only applies to the vector
layer so it should not be active when using the bitmap layer. The
pencil tool has a feather property, but it can't be changed, so
quick sizing should not be enabled for that.

Fixes #1186