Fix crash and incorrect color issues with eyedropper tool
I intially was fixing an issue where translucent colors would not
be picked correctly, and this was because it was not correctly
reversing the premultiply. I did some refactoring along with this
because there was lots of duplciated code between updateFrontColor
and pointerMoveEvent. When doing this, I noticed that there was
no check for getLast*ImageAtFrame returning null, which means that
Pencil2D will instantly crash with the eyedropper tool whenever
they current frame is before the first keyframe. All these issues
have been fixed.