Ignore mouse move event if not actively drawing a stroke
When using a Wacom tablet with Windows 7 compatibility mode, a move
event and release is sometimes generated after a release event in
the same stroke. We already aware of the second release event which
caused #677 and fixed that in #1188. However we were not aware of
the additional move event at the time, so the workaround was not
applied to that.

Fixes #1196