Enable redo action by default
Redo and undo are supposed to be enabled all the time when the edit
menu is closed so that shortcuts can be received without having to
change the state of the action every time the undo stack changes.
However the redo action wasn't being enabled until the edit menu
was closed.