Optimize flood fill of when expanding the target image
This treats everything outside the minimally bounded target image
as a single region (which it is). compareColor only needs to be
called once, and if it's true, then all of the transparent "border"
region can be filled at once. Ideally this would be done with a
rectangular fill with a mask when creating replaceImage, but
unforunately that information needs to be present in filledPixels
if expandFill is run. Thankfull this is pretty fast anyway.