UI Feature: Improve toolbox flow layout (#1916)
* FlowLayout: rewrite for maintainability sake

* Introduce customised flow layout for toolbox

Because I would like to change the alignment of the items on the last line

Also fixed a bunch of layout spacing inconsistencies

* Fix center alignment issue being off by one item

* Move Toolbox into own widget for more control of sizing

* Fix dock widget expansion caused by bucket and camera tools being layout

This happens because they will appear next to each other initially but then hidden away. The consequence of that
is the dockwidget sizes itself based on that, even if it they are removed right after.

* Add missing license header

* Cleanup